6-phase backend refactoring: security, auth consolidation, arch cleanup, dead code removal, AdminRepository port split, performance
