--- Goal.md
@@ -1,60 +1,0 @@
-# Goal
-
-## Why am I building this project?
-
-An actual tool for university students to practice coding — free, accessible, and comprehensive.
-
-## What is the ultimate purpose?
-
-A free, open-source LeetCode alternative with built-in AI coaching, evolving into a full CS curriculum platform (DSA, Programming Languages, DBMS, OOP, Web Dev) that professors recommend to their classes.
-
----
-
-## What problem does this website solve?
-
-University students lack a free, structured platform to practice coding interview questions AND learn programming languages (C, Python, Java) with instant AI coaching — without paying for LeetCode Premium or Hackerrank licenses.
-
-## Who is it for?
-
-- **Struggling students** who need hand-holding through the basics
-- **Interview grinders** preparing for tech internships and jobs
-- **Curious non-CS majors** who want to learn programming
-- **Professors** who need a free, curriculum-aligned tool to recommend
-
-All majors welcome. Multi-institution, drop-in, voluntary.
-
-## What makes it different from alternatives?
-
-|                         | CodeCoach AI                        | LeetCode | Hackerrank      | Exercism            |
-| ----------------------- | ----------------------------------- | -------- | --------------- | ------------------- |
-| **Free**                | Yes (BYO API key)                   | Freemium | Freemium        | Free                |
-| **AI Coaching**         | Built-in (NVIDIA NIM)               | No       | Limited         | No                  |
-| **Language Curriculum** | C, Python, Java lessons + exercises | No       | Limited tracks  | Yes, exercises only |
-| **Open Source**         | Yes                                 | No       | No              | Yes                 |
-| **Professor-Ready**     | Curriculum mapping, privacy-first   | No       | Enterprise only | No                  |
-
-## In one sentence, what does this website do?
-
-A free AI-powered coding platform where university students practice DSA problems and learn programming languages through structured lessons with real-time coaching — no payment needed.
-
-## Scope
-
-### Phase 1 — DSA Practice (36/100 questions generated)
-
-- 100 coding questions (30 Easy / 50 Medium / 20 Hard) across 14 standard topics
-- AI coaching with BYO NVIDIA API key
-- Email/password + Google OAuth (Supabase)
-
-### Phase 2 — Programming Language Curriculum ✅ (Complete)
-
-- C, Python, Java — 3 courses, 9 modules, 27 lessons (18 theory + 9 exercises)
-- Context-aware AI coaching per lesson
-- `/learn` navigation
-
-### Phase 3 — Future Modules
-
-- DBMS / SQL
-- OOP & Design Patterns
-- Web Development (React, Node)
-- Theory / MCQ question type
-- Classroom dashboard
