+++ backend/app/api/__init__.py
@@ -1,4 +1,27 @@
-# API package initialization
-from . import coach, run, questions, submit, health, debug
+from . import (
+    coach,
+    run,
+    questions,
+    submit,
+    health,
+    debug,
+    auth,
+    courses,
+    progress,
+    admin,
+    question_validation,
+)
 
-__all__ = ["coach", "run", "questions", "submit", "health", "debug"]
+__all__ = [
+    "coach",
+    "run",
+    "questions",
+    "submit",
+    "health",
+    "debug",
+    "auth",
+    "courses",
+    "progress",
+    "admin",
+    "question_validation",
+]
