+++ backend/app/api/courses.py
@@ -6,7 +6,7 @@ from app.ports.course_repository import CourseRepository
 from app.ports.progress_repository import ProgressRepository
 from app.services.course_service import CourseService
 from app.services.redis_service import RedisCache
-from app.api.auth import get_optional_current_user
+from app.api.auth_deps import get_optional_current_user
 from app.api.dependencies import get_course_repo, get_progress_repo, get_redis_cache
 from app.models.auth_schemas import UserResponse
