+++ backend/app/api/questions.py
@@ -2,19 +2,10 @@ from fastapi import APIRouter, Depends, HTTPException, Query
 from typing import Optional
 
 from app.models.schemas import Question, QuestionsListResponse, Difficulty
-from app.ports.question_repository import QuestionRepository
-from app.services.questions_service import QuestionsService
-from app.services.redis_service import RedisCache
-from app.api.dependencies import get_question_repo, get_redis_cache
+from app.services.question_bank import QuestionBank, QuestionFilters
+from app.api.dependencies import get_question_bank
 
 router = APIRouter()
-
-
-def get_questions_service(
-    question_repo: QuestionRepository = Depends(get_question_repo),
-    cache: Optional[RedisCache] = Depends(get_redis_cache),
-) -> QuestionsService:
-    return QuestionsService(repository=question_repo, cache=cache)
 
 
 @router.get("/", response_model=QuestionsListResponse)
@@ -23,38 +14,17 @@ async def get_questions(
     category: Optional[str] = Query(None, description="Filter by category"),
     page: int = Query(1, ge=1, description="Page number"),
     per_page: int = Query(100, ge=1, le=100, description="Items per page"),
-    questions_service: QuestionsService = Depends(get_questions_service),
+    bank: QuestionBank = Depends(get_question_bank),
 ):
-    """
-    Get all questions with optional filtering and pagination.
-
-    Supports filtering by difficulty and category, with pagination.
-    """
-
     try:
-        questions = await questions_service.get_all_questions(
-            difficulty=difficulty, category=category, page=page, per_page=per_page
-        )
-
-        total = await questions_service.get_total_count()
-
-        # Apply filtering to total count
-        if difficulty or category:
-            filtered_total = len(
-                await questions_service.get_all_questions(
-                    difficulty=difficulty,
-                    category=category,
-                    page=1,
-                    per_page=10000,  # Get all for count
-                )
+        result = await bank.query(
+            QuestionFilters(
+                difficulty=difficulty, category=category, page=page, per_page=per_page
             )
-        else:
-            filtered_total = total
-
-        return QuestionsListResponse(
-            questions=questions, total=filtered_total, page=page, per_page=per_page
         )
-
+        return QuestionsListResponse(
+            questions=result.items, total=result.total, page=page, per_page=per_page
+        )
     except Exception as e:
         raise HTTPException(
             status_code=500, detail=f"Error fetching questions: {str(e)}"
@@ -63,14 +33,11 @@ async def get_questions(
 
 @router.get("/categories")
 async def get_categories(
-    questions_service: QuestionsService = Depends(get_questions_service),
+    bank: QuestionBank = Depends(get_question_bank),
 ):
-    """Get all available question categories."""
-
     try:
-        categories = await questions_service.get_categories()
-        return {"categories": categories}
-
+        stats = await bank.stats()
+        return {"categories": stats.categories}
     except Exception as e:
         raise HTTPException(
             status_code=500, detail=f"Error fetching categories: {str(e)}"
@@ -79,14 +46,11 @@ async def get_categories(
 
 @router.get("/companies")
 async def get_companies(
-    questions_service: QuestionsService = Depends(get_questions_service),
+    bank: QuestionBank = Depends(get_question_bank),
 ):
-    """Get all company tags used in questions."""
-
     try:
-        companies = await questions_service.get_company_tags()
-        return {"companies": companies}
-
+        stats = await bank.stats()
+        return {"companies": stats.companies}
     except Exception as e:
         raise HTTPException(
             status_code=500, detail=f"Error fetching companies: {str(e)}"
@@ -100,34 +64,28 @@ async def search_questions(
     category: Optional[str] = Query(None, description="Filter by category"),
     page: int = Query(1, ge=1, description="Page number"),
     per_page: int = Query(100, ge=1, le=100, description="Items per page"),
-    questions_service: QuestionsService = Depends(get_questions_service),
+    bank: QuestionBank = Depends(get_question_bank),
 ):
-    """
-    Search questions by title or description.
-
-    Supports text search with optional difficulty and category filters.
-    """
-
     try:
         if not q.strip():
             raise HTTPException(status_code=400, detail="Search query cannot be empty")
 
-        all_results = await questions_service.search_questions(
-            query=q, difficulty=difficulty, category=category
+        result = await bank.query(
+            QuestionFilters(
+                query=q, difficulty=difficulty, category=category, per_page=10000
+            )
         )
 
-        # Pagination
         start_idx = (page - 1) * per_page
         end_idx = start_idx + per_page
-        paginated_results = all_results[start_idx:end_idx]
+        paginated = result.items[start_idx:end_idx]
 
         return QuestionsListResponse(
-            questions=paginated_results,
-            total=len(all_results),
+            questions=paginated,
+            total=result.total,
             page=page,
             per_page=per_page,
         )
-
     except HTTPException as he:
         raise he
     except Exception as e:
@@ -138,23 +96,17 @@ async def search_questions(
 
 @router.get("/stats")
 async def get_question_stats(
-    questions_service: QuestionsService = Depends(get_questions_service),
+    bank: QuestionBank = Depends(get_question_bank),
 ):
-    """Get statistics about the question bank."""
-
     try:
-        total = await questions_service.get_total_count()
-        difficulty_counts = await questions_service.get_difficulty_counts()
-        category_counts = await questions_service.get_category_counts()
-
+        stats = await bank.stats()
         return {
-            "total": total,
-            "difficulty_counts": difficulty_counts,
-            "category_counts": category_counts,
-            "categories": await questions_service.get_categories(),
-            "companies": await questions_service.get_company_tags(),
+            "total": stats.total,
+            "difficulty_counts": stats.difficulty_counts,
+            "category_counts": stats.category_counts,
+            "categories": stats.categories,
+            "companies": stats.companies,
         }
-
     except Exception as e:
         raise HTTPException(
             status_code=500, detail=f"Error fetching statistics: {str(e)}"
@@ -166,25 +118,21 @@ async def get_questions_by_category(
     category: str,
     page: int = Query(1, ge=1, description="Page number"),
     per_page: int = Query(100, ge=1, le=100, description="Items per page"),
-    questions_service: QuestionsService = Depends(get_questions_service),
+    bank: QuestionBank = Depends(get_question_bank),
 ):
-    """Get questions filtered by category."""
-
     try:
-        questions = await questions_service.get_questions_by_category(category)
+        result = await bank.query(QuestionFilters(category=category, per_page=10000))
 
-        # Pagination
         start_idx = (page - 1) * per_page
         end_idx = start_idx + per_page
-        paginated_questions = questions[start_idx:end_idx]
+        paginated = result.items[start_idx:end_idx]
 
         return QuestionsListResponse(
-            questions=paginated_questions,
-            total=len(questions),
+            questions=paginated,
+            total=result.total,
             page=page,
             per_page=per_page,
         )
-
     except Exception as e:
         raise HTTPException(
             status_code=500, detail=f"Error fetching questions by category: {str(e)}"
@@ -196,25 +144,23 @@ async def get_questions_by_difficulty(
     difficulty: Difficulty,
     page: int = Query(1, ge=1, description="Page number"),
     per_page: int = Query(100, ge=1, le=100, description="Items per page"),
-    questions_service: QuestionsService = Depends(get_questions_service),
+    bank: QuestionBank = Depends(get_question_bank),
 ):
-    """Get questions filtered by difficulty."""
-
     try:
-        questions = await questions_service.get_questions_by_difficulty(difficulty)
+        result = await bank.query(
+            QuestionFilters(difficulty=difficulty, per_page=10000)
+        )
 
-        # Pagination
         start_idx = (page - 1) * per_page
         end_idx = start_idx + per_page
-        paginated_questions = questions[start_idx:end_idx]
+        paginated = result.items[start_idx:end_idx]
 
         return QuestionsListResponse(
-            questions=paginated_questions,
-            total=len(questions),
+            questions=paginated,
+            total=result.total,
             page=page,
             per_page=per_page,
         )
-
     except Exception as e:
         raise HTTPException(
             status_code=500, detail=f"Error fetching questions by difficulty: {str(e)}"
@@ -224,18 +170,10 @@ async def get_questions_by_difficulty(
 @router.get("/{question_id}", response_model=Question)
 async def get_question(
     question_id: str,
-    questions_service: QuestionsService = Depends(get_questions_service),
+    bank: QuestionBank = Depends(get_question_bank),
 ):
-    """
-    Get a specific question by ID.
-
-    Returns full question details including description, examples, test cases, and starter code.
-    """
-
     try:
-        question = await questions_service.get_question_by_id(question_id)
-        return question
-
+        return await bank.get(question_id)
     except HTTPException as he:
         raise he
     except Exception as e:
