+++ backend/app/ports/course_repository.py
@@ -22,3 +22,9 @@ class CourseRepository(ABC):
 
     @abstractmethod
     async def get_modules_by_course(self, course_id: str) -> List[Module]: ...
+
+    async def get_modules_by_course_batch(self, course_ids: List[str]) -> List[Module]:
+        result = []
+        for cid in course_ids:
+            result.extend(await self.get_modules_by_course(cid))
+        return result
