+++ backend/tests/unit/test_question_validation_use_cases.py
@@ -6,20 +6,16 @@ are made available to users. Tests focus on Python coverage using Piston.
 """
 
 import pytest
-from typing import Dict, Any, List
-from unittest.mock import AsyncMock, MagicMock, patch
-import asyncio
+from typing import Dict, Any
+from unittest.mock import AsyncMock
 import json
 import os
 import tempfile
 
-from app.models.schemas import Question, TestCase, Example, StarterCode, Difficulty
+from app.models.schemas import Question, Difficulty
 from app.models.question_validation_schemas import (
     ValidationUseCase,
     ValidationSeverity,
-    ValidationIssue,
-    UseCaseValidationResult,
-    QuestionValidationResult,
     QuestionValidationConfig,
 )
 
@@ -27,6 +23,7 @@ from app.models.question_validation_schemas import (
 # ============================================================================
 # Test Fixtures
 # ============================================================================
+
 
 @pytest.fixture
 def valid_question_data() -> Dict[str, Any]:
@@ -41,13 +38,13 @@ def valid_question_data() -> Dict[str, Any]:
         "starter": {
             "python": "def solve(nums: List[int]) -> List[int]:\n    # Your code here\n    pass",
             "javascript": "function solve(nums) {\n    // Your code here\n}",
-            "java": "class Solution {\n    public int[] solve(int[] nums) {\n        // Your code here\n    }\n}"
+            "java": "class Solution {\n    public int[] solve(int[] nums) {\n        // Your code here\n    }\n}",
         },
         "examples": [
             {
                 "input": "nums = [1,2,3]",
                 "output": "[1,2,3]",
-                "explanation": "Test explanation"
+                "explanation": "Test explanation",
             }
         ],
         "test_cases": [
@@ -55,20 +52,20 @@ def valid_question_data() -> Dict[str, Any]:
                 "input": "[1,2,3]",
                 "expected_output": "[1,2,3]",
                 "description": "Basic test case",
-                "hidden": False
+                "hidden": False,
             },
             {
                 "input": "[4,5,6]",
                 "expected_output": "[4,5,6]",
                 "description": "Another test case",
-                "hidden": True
-            }
+                "hidden": True,
+            },
         ],
         "hints": ["Think about the problem"],
         "solution": "Return the array as is.",
         "time_complexity": "O(n)",
         "space_complexity": "O(1)",
-        "constraints": ["1 <= nums.length <= 100"]
+        "constraints": ["1 <= nums.length <= 100"],
     }
 
 
@@ -91,12 +88,12 @@ def invalid_question_missing_id() -> Question:
         "starter": {
             "python": "def solve(): pass",
             "javascript": "function solve() {}",
-            "java": "class Solution { public void solve() {} }"
+            "java": "class Solution { public void solve() {} }",
         },
         "examples": [{"input": "test", "output": "test"}],
         "test_cases": [{"input": "test", "expected_output": "test"}],
         "hints": [],
-        "constraints": []
+        "constraints": [],
     }
     return Question(**data)
 
@@ -114,12 +111,12 @@ def question_with_invalid_test_cases() -> Question:
         "starter": {
             "python": "def solve(): pass",
             "javascript": "function solve() {}",
-            "java": "class Solution { public void solve() {} }"
+            "java": "class Solution { public void solve() {} }",
         },
         "examples": [{"input": "test", "output": "test"}],
         "test_cases": [],  # Invalid: no test cases
         "hints": [],
-        "constraints": []
+        "constraints": [],
     }
     return Question(**data)
 
@@ -137,12 +134,12 @@ def question_with_bad_starter_code() -> Question:
         "starter": {
             "python": "def solve(:",  # Invalid Python syntax
             "javascript": "function solve() {",
-            "java": "class Solution { public void solve() {} }"
+            "java": "class Solution { public void solve() {} }",
         },
         "examples": [{"input": "test", "output": "test"}],
         "test_cases": [{"input": "test", "expected_output": "test"}],
         "hints": [],
-        "constraints": []
+        "constraints": [],
     }
     return Question(**data)
 
@@ -151,7 +148,7 @@ def question_with_bad_starter_code() -> Question:
 def mock_piston_service():
     """Create a mock executor for testing."""
     from app.ports.code_executor import CodeExecutor
-    from unittest.mock import AsyncMock
+
     return AsyncMock(spec=CodeExecutor)
 
 
@@ -159,71 +156,80 @@ def mock_piston_service():
 # StructureValidationUseCase Tests
 # ============================================================================
 
+
 class TestStructureValidationUseCase:
     """Tests for the structure validation use case."""
-    
+
     @pytest.mark.asyncio
     async def test_valid_question_passes_structure_validation(self, valid_question):
         """Test that a valid question passes structure validation."""
         from app.services.question_validator import StructureValidationUseCase
-        
+
         use_case = StructureValidationUseCase()
         result = await use_case.execute(valid_question)
-        
+
         assert result.passed is True
         assert result.use_case == ValidationUseCase.STRUCTURE
         assert len(result.issues) == 0
-    
+
     @pytest.mark.asyncio
-    async def test_question_with_empty_id_fails_validation(self, invalid_question_missing_id):
+    async def test_question_with_empty_id_fails_validation(
+        self, invalid_question_missing_id
+    ):
         """Test that a question with empty ID fails validation."""
         from app.services.question_validator import StructureValidationUseCase
-        
+
         use_case = StructureValidationUseCase()
         result = await use_case.execute(invalid_question_missing_id)
-        
+
         assert result.passed is False
         assert any(issue.field == "id" for issue in result.issues)
-    
+
     @pytest.mark.asyncio
-    async def test_question_with_short_title_fails_validation(self, valid_question_data):
+    async def test_question_with_short_title_fails_validation(
+        self, valid_question_data
+    ):
         """Test that a question with too short title fails validation."""
         valid_question_data["title"] = "Hi"  # Too short
         question = Question(**valid_question_data)
-        
+
         from app.services.question_validator import StructureValidationUseCase
-        
+
         use_case = StructureValidationUseCase()
         result = await use_case.execute(question)
-        
+
         assert result.passed is False
         assert any(issue.field == "title" for issue in result.issues)
-    
+
     @pytest.mark.asyncio
-    async def test_question_with_short_description_fails_validation(self, valid_question_data):
+    async def test_question_with_short_description_fails_validation(
+        self, valid_question_data
+    ):
         """Test that a question with too short description fails validation."""
         valid_question_data["description"] = "Too short"  # Less than 50 chars
         question = Question(**valid_question_data)
-        
+
         from app.services.question_validator import StructureValidationUseCase
-        
+
         use_case = StructureValidationUseCase()
         result = await use_case.execute(question)
-        
+
         assert result.passed is False
         assert any(issue.field == "description" for issue in result.issues)
-    
+
     @pytest.mark.asyncio
-    async def test_question_missing_starter_language_fails_validation(self, valid_question_data):
+    async def test_question_missing_starter_language_fails_validation(
+        self, valid_question_data
+    ):
         """Test that missing starter code for a language fails validation."""
         valid_question_data["starter"]["python"] = ""  # Empty Python starter
         question = Question(**valid_question_data)
-        
+
         from app.services.question_validator import StructureValidationUseCase
-        
+
         use_case = StructureValidationUseCase()
         result = await use_case.execute(question)
-        
+
         assert result.passed is False
         assert any(issue.language == "python" for issue in result.issues)
 
@@ -232,99 +238,124 @@ class TestStructureValidationUseCase:
 # TestCaseValidationUseCase Tests
 # ============================================================================
 
+
 class TestTestCaseValidationUseCase:
     """Tests for test case validation use case."""
-    
+
     @pytest.mark.asyncio
-    async def test_question_with_no_test_cases_fails_validation(self, question_with_invalid_test_cases):
+    async def test_question_with_no_test_cases_fails_validation(
+        self, question_with_invalid_test_cases
+    ):
         """Test that a question with no test cases fails validation."""
         from app.services.question_validator import TestCaseValidationUseCase
-        
+
         use_case = TestCaseValidationUseCase()
         result = await use_case.execute(question_with_invalid_test_cases)
-        
+
         assert result.passed is False
-        assert any(issue.use_case == ValidationUseCase.TEST_CASES for issue in result.issues)
-    
+        assert any(
+            issue.use_case == ValidationUseCase.TEST_CASES for issue in result.issues
+        )
+
     @pytest.mark.asyncio
-    async def test_test_case_with_empty_input_fails_validation(self, valid_question_data):
+    async def test_test_case_with_empty_input_fails_validation(
+        self, valid_question_data
+    ):
         """Test that test case with empty input fails validation."""
         valid_question_data["test_cases"] = [
             {"input": "", "expected_output": "result"}  # Empty input
         ]
         question = Question(**valid_question_data)
-        
+
         from app.services.question_validator import TestCaseValidationUseCase
-        
+
         use_case = TestCaseValidationUseCase()
         result = await use_case.execute(question)
-        
+
         assert result.passed is False
-    
+
     @pytest.mark.asyncio
-    async def test_test_case_executability_validated_with_piston(self, valid_question, mock_piston_service):
+    async def test_test_case_executability_validated_with_piston(
+        self, valid_question, mock_piston_service
+    ):
         """Test that test case executability is validated using Piston."""
         from app.services.question_validator import TestCaseValidationUseCase
         from app.ports.code_executor import ExecutionResult
 
         # Mock executor response for successful execution
-        mock_piston_service.execute.return_value = ExecutionResult(stdout="[1, 2, 3]", exit_code=0)
+        mock_piston_service.execute.return_value = ExecutionResult(
+            stdout="[1, 2, 3]", exit_code=0
+        )
 
         use_case = TestCaseValidationUseCase(executor=mock_piston_service)
         result = await use_case.execute(valid_question)
 
         # Should have called executor to validate test case executability
         assert mock_piston_service.execute.called or result.passed is True
-    
+
     @pytest.mark.asyncio
     async def test_hidden_test_cases_have_different_inputs(self, valid_question_data):
         """Test that hidden test cases have different inputs from visible ones."""
         # Create duplicate test cases (one hidden, one visible with same input)
         valid_question_data["test_cases"] = [
             {"input": "[1,2,3]", "expected_output": "[1,2,3]", "hidden": False},
-            {"input": "[1,2,3]", "expected_output": "[1,2,3]", "hidden": True}  # Same input!
+            {
+                "input": "[1,2,3]",
+                "expected_output": "[1,2,3]",
+                "hidden": True,
+            },  # Same input!
         ]
         question = Question(**valid_question_data)
-        
+
         from app.services.question_validator import TestCaseValidationUseCase
-        
+
         use_case = TestCaseValidationUseCase()
         result = await use_case.execute(question)
-        
+
         # Should warn about duplicate test cases
-        assert any(issue.severity == ValidationSeverity.WARNING for issue in result.issues)
+        assert any(
+            issue.severity == ValidationSeverity.WARNING for issue in result.issues
+        )
 
 
 # ============================================================================
 # StarterCodeValidationUseCase Tests
 # ============================================================================
 
+
 class TestStarterCodeValidationUseCase:
     """Tests for starter code validation use case."""
-    
+
     @pytest.mark.asyncio
-    async def test_valid_starter_code_passes_validation(self, valid_question, mock_piston_service):
+    async def test_valid_starter_code_passes_validation(
+        self, valid_question, mock_piston_service
+    ):
         """Test that valid starter code passes validation."""
         from app.services.question_validator import StarterCodeValidationUseCase
         from app.ports.code_executor import ExecutionResult
 
         # Mock executor to return success for syntax check
-        mock_piston_service.execute.return_value = ExecutionResult(stdout="", exit_code=0)
+        mock_piston_service.execute.return_value = ExecutionResult(
+            stdout="", exit_code=0
+        )
 
         use_case = StarterCodeValidationUseCase(executor=mock_piston_service)
         result = await use_case.execute(valid_question)
 
         assert result.passed is True
-    
+
     @pytest.mark.asyncio
-    async def test_invalid_python_starter_code_fails_validation(self, question_with_bad_starter_code):
+    async def test_invalid_python_starter_code_fails_validation(
+        self, question_with_bad_starter_code
+    ):
         """Test that invalid Python starter code fails validation."""
         from app.services.question_validator import StarterCodeValidationUseCase
-        from unittest.mock import AsyncMock
         from app.ports.code_executor import CodeExecutor, ExecutionResult
 
         executor = AsyncMock(spec=CodeExecutor)
-        executor.execute.return_value = ExecutionResult(exit_code=1, stderr="SyntaxError: invalid syntax")
+        executor.execute.return_value = ExecutionResult(
+            exit_code=1, stderr="SyntaxError: invalid syntax"
+        )
 
         use_case = StarterCodeValidationUseCase(executor=executor)
         result = await use_case.execute(question_with_bad_starter_code)
@@ -336,14 +367,13 @@ class TestStarterCodeValidationUseCase:
     async def test_all_languages_validated(self, valid_question):
         """Test that all three languages are validated."""
         from app.services.question_validator import StarterCodeValidationUseCase
-        from unittest.mock import AsyncMock
         from app.ports.code_executor import CodeExecutor, ExecutionResult
 
         executor = AsyncMock(spec=CodeExecutor)
         executor.execute.return_value = ExecutionResult(stdout="", exit_code=0)
 
         use_case = StarterCodeValidationUseCase(executor=executor)
-        result = await use_case.execute(valid_question)
+        await use_case.execute(valid_question)
 
         # Should validate Python, JavaScript, and Java
         assert executor.execute.call_count >= 3
@@ -353,17 +383,20 @@ class TestStarterCodeValidationUseCase:
 # SolutionValidationUseCase Tests
 # ============================================================================
 
+
 class TestSolutionValidationUseCase:
     """Tests for solution validation use case."""
-    
+
     @pytest.mark.asyncio
-    async def test_solution_passes_all_test_cases(self, valid_question_data, mock_piston_service):
+    async def test_solution_passes_all_test_cases(
+        self, valid_question_data, mock_piston_service
+    ):
         """Test that the reference solution passes all test cases."""
         from app.services.question_validator import SolutionValidationUseCase
         from app.ports.code_executor import ExecutionResult
 
         # Create a question with actual solution code
-        valid_question_data["starter"]["python"] = '''
+        valid_question_data["starter"]["python"] = """
 def solve(nums):
     return nums
 
@@ -373,7 +406,7 @@ lines = sys.stdin.read().strip().split('\\n')
 nums = json.loads(lines[0])
 result = solve(nums)
 print(json.dumps(result))
-'''
+"""
         valid_question_data["solution"] = "return nums"
         # Use single test case for simplicity
         valid_question_data["test_cases"] = [
@@ -381,22 +414,26 @@ print(json.dumps(result))
                 "input": "[1,2,3]",
                 "expected_output": "[1,2,3]",
                 "description": "Basic test case",
-                "hidden": False
+                "hidden": False,
             }
         ]
         question = Question(**valid_question_data)
 
         # Mock executor to return correct output
-        mock_piston_service.execute.return_value = ExecutionResult(stdout="[1,2,3]", exit_code=0)
+        mock_piston_service.execute.return_value = ExecutionResult(
+            stdout="[1,2,3]", exit_code=0
+        )
 
         use_case = SolutionValidationUseCase(executor=mock_piston_service)
         result = await use_case.execute(question)
 
         # Should pass since we have executable code
         assert result.passed is True
-    
+
     @pytest.mark.asyncio
-    async def test_solution_fails_test_case(self, valid_question_data, mock_piston_service):
+    async def test_solution_fails_test_case(
+        self, valid_question_data, mock_piston_service
+    ):
         """Test that solution failing a test case is detected."""
         # Create question with solution that doesn't match expected output
         valid_question_data["solution"] = "def solve(nums): return []"  # Wrong solution
@@ -406,25 +443,31 @@ print(json.dumps(result))
         from app.ports.code_executor import ExecutionResult
 
         # Mock executor to return wrong output
-        mock_piston_service.execute.return_value = ExecutionResult(stdout="[]", exit_code=0)
+        mock_piston_service.execute.return_value = ExecutionResult(
+            stdout="[]", exit_code=0
+        )
 
         use_case = SolutionValidationUseCase(executor=mock_piston_service)
         result = await use_case.execute(question)
 
         assert result.passed is False
-        assert any(issue.use_case == ValidationUseCase.SOLUTION for issue in result.issues)
-    
+        assert any(
+            issue.use_case == ValidationUseCase.SOLUTION for issue in result.issues
+        )
+
     @pytest.mark.asyncio
-    async def test_solution_missing_fails_validation(self, valid_question_data, mock_piston_service):
+    async def test_solution_missing_fails_validation(
+        self, valid_question_data, mock_piston_service
+    ):
         """Test that missing solution fails validation."""
         valid_question_data["solution"] = None
         question = Question(**valid_question_data)
-        
+
         from app.services.question_validator import SolutionValidationUseCase
-        
+
         use_case = SolutionValidationUseCase(executor=mock_piston_service)
         result = await use_case.execute(question)
-        
+
         assert result.passed is False
         assert any("solution" in issue.message.lower() for issue in result.issues)
 
@@ -433,109 +476,117 @@ print(json.dumps(result))
 # TimeLimitValidationUseCase Tests
 # ============================================================================
 
+
 class TestTimeLimitValidationUseCase:
     """Tests for time limit validation use case."""
-    
+
     @pytest.mark.asyncio
     async def test_default_time_limits_are_valid(self, valid_question):
         """Test that default time limits are valid."""
         from app.services.question_validator import TimeLimitValidationUseCase
-        
+
         use_case = TimeLimitValidationUseCase()
         result = await use_case.execute(valid_question)
-        
+
         assert result.passed is True
-    
+
     @pytest.mark.asyncio
     async def test_time_limit_within_bounds(self, valid_question_data):
         """Test that time limits within bounds pass validation."""
         # Time complexity implies reasonable time limit
         valid_question_data["time_complexity"] = "O(n)"
         question = Question(**valid_question_data)
-        
+
         from app.services.question_validator import TimeLimitValidationUseCase
-        
+
         use_case = TimeLimitValidationUseCase()
         result = await use_case.execute(question)
-        
+
         assert result.passed is True
-    
+
     @pytest.mark.asyncio
     async def test_slow_algorithm_warns_about_time_limit(self, valid_question_data):
         """Test that slow algorithms get warnings about time limits."""
         valid_question_data["time_complexity"] = "O(n^3)"
         question = Question(**valid_question_data)
-        
+
         from app.services.question_validator import TimeLimitValidationUseCase
-        
+
         use_case = TimeLimitValidationUseCase()
         result = await use_case.execute(question)
-        
+
         # Should warn about potential time limit issues
-        assert any(issue.severity == ValidationSeverity.WARNING for issue in result.issues)
+        assert any(
+            issue.severity == ValidationSeverity.WARNING for issue in result.issues
+        )
 
 
 # ============================================================================
 # FunctionSignatureValidationUseCase Tests
 # ============================================================================
 
+
 class TestFunctionSignatureValidationUseCase:
     """Tests for function signature validation use case."""
-    
+
     @pytest.mark.asyncio
     async def test_valid_function_signature_passes(self, valid_question_data):
         """Test that valid function signature passes validation."""
         from app.services.question_validator import FunctionSignatureValidationUseCase
-        
+
         # Create question with valid Java method signature and Python type hints
-        valid_question_data["starter"]["java"] = '''
+        valid_question_data["starter"]["java"] = """
 class Solution {
     public int[] solve(int[] nums) {
         // Your code here
         return new int[0];
     }
 }
-'''
-        valid_question_data["starter"]["python"] = '''
+"""
+        valid_question_data["starter"]["python"] = """
 from typing import List
 
 def solve(nums: List[int]) -> List[int]:
     # Your code here
     pass
-'''
+"""
         question = Question(**valid_question_data)
-        
+
         use_case = FunctionSignatureValidationUseCase()
         result = await use_case.execute(question)
-        
+
         assert result.passed is True
-    
+
     @pytest.mark.asyncio
     async def test_missing_type_hints_warning(self, valid_question_data):
         """Test that missing type hints generate warnings."""
         # Remove type hints from Python starter
         valid_question_data["starter"]["python"] = "def solve(nums):\n    pass"
         question = Question(**valid_question_data)
-        
+
         from app.services.question_validator import FunctionSignatureValidationUseCase
-        
+
         use_case = FunctionSignatureValidationUseCase(require_type_hints=True)
         result = await use_case.execute(question)
-        
-        assert any(issue.severity == ValidationSeverity.WARNING for issue in result.issues)
-    
+
+        assert any(
+            issue.severity == ValidationSeverity.WARNING for issue in result.issues
+        )
+
     @pytest.mark.asyncio
     async def test_invalid_return_type_fails(self, valid_question_data):
         """Test that invalid return type fails validation."""
         # Use invalid return type
-        valid_question_data["starter"]["python"] = "def solve(nums) -> InvalidType:\n    pass"
+        valid_question_data["starter"]["python"] = (
+            "def solve(nums) -> InvalidType:\n    pass"
+        )
         question = Question(**valid_question_data)
-        
+
         from app.services.question_validator import FunctionSignatureValidationUseCase
-        
+
         use_case = FunctionSignatureValidationUseCase()
         result = await use_case.execute(question)
-        
+
         # Should warn about potentially invalid type hint
         assert len(result.issues) > 0
 
@@ -544,34 +595,35 @@ def solve(nums: List[int]) -> List[int]:
 # OutputFormatValidationUseCase Tests
 # ============================================================================
 
+
 class TestOutputFormatValidationUseCase:
     """Tests for output format validation use case."""
-    
+
     @pytest.mark.asyncio
     async def test_valid_output_format_passes(self, valid_question):
         """Test that valid output format passes validation."""
         from app.services.question_validator import OutputFormatValidationUseCase
-        
+
         use_case = OutputFormatValidationUseCase()
         result = await use_case.execute(valid_question)
-        
+
         assert result.passed is True
-    
+
     @pytest.mark.asyncio
     async def test_inconsistent_output_format_fails(self, valid_question_data):
         """Test that inconsistent output formats across test cases fail."""
         # Different output formats for same problem
         valid_question_data["test_cases"] = [
             {"input": "test1", "expected_output": "[1,2,3]"},  # JSON array
-            {"input": "test2", "expected_output": "hello"}     # String - inconsistent!
+            {"input": "test2", "expected_output": "hello"},  # String - inconsistent!
         ]
         question = Question(**valid_question_data)
-        
+
         from app.services.question_validator import OutputFormatValidationUseCase
-        
+
         use_case = OutputFormatValidationUseCase()
         result = await use_case.execute(question)
-        
+
         assert result.passed is False
 
 
@@ -579,24 +631,27 @@ class TestOutputFormatValidationUseCase:
 # QuestionValidatorService Tests
 # ============================================================================
 
+
 class TestQuestionValidatorService:
     """Tests for the main question validator service."""
-    
+
     @pytest.mark.asyncio
-    async def test_valid_question_passes_all_validations(self, valid_question_data, mock_piston_service):
+    async def test_valid_question_passes_all_validations(
+        self, valid_question_data, mock_piston_service
+    ):
         """Test that a valid question passes all validations."""
         from app.services.question_validator import QuestionValidatorService
         from app.ports.code_executor import ExecutionResult
 
         # Create question with valid Java method signature and executable solution
-        valid_question_data["starter"]["java"] = '''
+        valid_question_data["starter"]["java"] = """
 class Solution {
     public int[] solve(int[] nums) {
         return nums;
     }
 }
-'''
-        valid_question_data["starter"]["python"] = '''
+"""
+        valid_question_data["starter"]["python"] = """
 from typing import List
 
 def solve(nums: List[int]) -> List[int]:
@@ -608,21 +663,21 @@ lines = sys.stdin.read().strip().split('\\n')
 nums = json.loads(lines[0])
 result = solve(nums)
 print(json.dumps(result))
-'''
+"""
         valid_question_data["solution"] = "return nums"
         valid_question_data["test_cases"] = [
             {
                 "input": "[1,2,3]",
                 "expected_output": "[1,2,3]",
                 "description": "Basic test case",
-                "hidden": False
+                "hidden": False,
             },
             {
                 "input": "[4,5,6]",
                 "expected_output": "[4,5,6]",
                 "description": "Another test case",
-                "hidden": True
-            }
+                "hidden": True,
+            },
         ]
         question = Question(**valid_question_data)
 
@@ -630,7 +685,7 @@ print(json.dumps(result))
 
         # Mock executor to return appropriate output based on input
         def mock_execute_side_effect(language, code, stdin="", version=None):
-            if '[4,5,6]' in stdin:
+            if "[4,5,6]" in stdin:
                 return ExecutionResult(stdout="[4,5,6]", exit_code=0)
             return ExecutionResult(stdout="[1,2,3]", exit_code=0)
 
@@ -640,30 +695,38 @@ print(json.dumps(result))
 
         assert result.valid is True
         assert result.error_count == 0
-    
+
     @pytest.mark.asyncio
-    async def test_invalid_question_fails_validation(self, question_with_invalid_test_cases, mock_piston_service):
+    async def test_invalid_question_fails_validation(
+        self, question_with_invalid_test_cases, mock_piston_service
+    ):
         """Test that an invalid question fails validation."""
         from app.services.question_validator import QuestionValidatorService
         from app.ports.code_executor import ExecutionResult
 
-        mock_piston_service.execute.return_value = ExecutionResult(stdout="", exit_code=0)
+        mock_piston_service.execute.return_value = ExecutionResult(
+            stdout="", exit_code=0
+        )
 
         service = QuestionValidatorService(executor=mock_piston_service)
         result = await service.validate_question(question_with_invalid_test_cases)
 
         assert result.valid is False
         assert result.error_count > 0
-    
+
     @pytest.mark.asyncio
-    async def test_all_use_cases_are_executed(self, valid_question, mock_piston_service):
+    async def test_all_use_cases_are_executed(
+        self, valid_question, mock_piston_service
+    ):
         """Test that all validation use cases are executed."""
         from app.services.question_validator import QuestionValidatorService
         from app.ports.code_executor import ExecutionResult
 
         service = QuestionValidatorService(executor=mock_piston_service)
 
-        mock_piston_service.execute.return_value = ExecutionResult(stdout="result", exit_code=0)
+        mock_piston_service.execute.return_value = ExecutionResult(
+            stdout="result", exit_code=0
+        )
 
         result = await service.validate_question(valid_question)
 
@@ -680,24 +743,22 @@ print(json.dumps(result))
 
         for use_case in expected_use_cases:
             assert use_case in result.results
-    
+
     @pytest.mark.asyncio
-    async def test_validation_can_be_skipped_for_specific_use_cases(self, valid_question, mock_piston_service):
+    async def test_validation_can_be_skipped_for_specific_use_cases(
+        self, valid_question, mock_piston_service
+    ):
         """Test that specific use cases can be skipped."""
         from app.services.question_validator import QuestionValidatorService
-        from app.models.question_validation_schemas import QuestionValidationConfig
         from app.ports.code_executor import ExecutionResult
 
-        config = QuestionValidationConfig(
-            skip_use_cases=[ValidationUseCase.SOLUTION]
-        )
+        config = QuestionValidationConfig(skip_use_cases=[ValidationUseCase.SOLUTION])
 
-        service = QuestionValidatorService(
-            executor=mock_piston_service,
-            config=config
-        )
+        service = QuestionValidatorService(executor=mock_piston_service, config=config)
 
-        mock_piston_service.execute.return_value = ExecutionResult(stdout="result", exit_code=0)
+        mock_piston_service.execute.return_value = ExecutionResult(
+            stdout="result", exit_code=0
+        )
 
         result = await service.validate_question(valid_question)
 
@@ -706,15 +767,17 @@ print(json.dumps(result))
 
 
 # ============================================================================
-# Integration with QuestionsService Tests
+# Integration with QuestionBank Tests
 # ============================================================================
 
-class TestQuestionsServiceValidationIntegration:
-    """Tests for validation integration with QuestionsService."""
-    
+
+class TestQuestionBankValidationIntegration:
+    """Tests for validation integration with QuestionBank."""
+
     @pytest.fixture
     def empty_repo(self):
         from app.repositories.file_question_repository import FileQuestionRepository
+
         fd, path = tempfile.mkstemp(suffix=".json")
         with os.fdopen(fd, "w", encoding="utf-8") as f:
             json.dump([], f)
@@ -722,30 +785,36 @@ class TestQuestionsServiceValidationIntegration:
         os.unlink(path)
 
     @pytest.mark.asyncio
-    async def test_invalid_question_not_loaded(self, question_with_invalid_test_cases, empty_repo):
+    async def test_invalid_question_not_loaded(
+        self, question_with_invalid_test_cases, empty_repo
+    ):
         """Test that invalid questions are not loaded into the question bank."""
-        from app.services.questions_service import QuestionsService
+        from app.services.question_bank import QuestionBank
 
-        service = QuestionsService(repository=empty_repo)
+        bank = QuestionBank(repository=empty_repo)
 
-        result = await service.add_question(question_with_invalid_test_cases, validate=False)
+        result = await bank.add(question_with_invalid_test_cases, validate=False)
 
         assert result.is_validated is False
 
     @pytest.mark.asyncio
-    async def test_validation_status_stored_with_question(self, valid_question, mock_piston_service, empty_repo):
+    async def test_validation_status_stored_with_question(
+        self, valid_question, mock_piston_service, empty_repo
+    ):
         """Test that validation status is stored with the question."""
-        from app.services.questions_service import QuestionsService
+        from app.services.question_bank import QuestionBank
         from app.services.question_validator import QuestionValidatorService
         from app.ports.code_executor import ExecutionResult
 
         validator = QuestionValidatorService(executor=mock_piston_service)
-        service = QuestionsService(repository=empty_repo, validator=validator)
+        bank = QuestionBank(repository=empty_repo, validator=validator)
 
         # Mock executor response for tests that execute code
-        mock_piston_service.execute.return_value = ExecutionResult(stdout="result", exit_code=0)
+        mock_piston_service.execute.return_value = ExecutionResult(
+            stdout="result", exit_code=0
+        )
 
-        result = await service.add_question(valid_question, validate=True)
+        await bank.add(valid_question, validate=True)
 
         statuses = await empty_repo.get_validation_statuses()
         assert valid_question.id in statuses
@@ -755,43 +824,46 @@ class TestQuestionsServiceValidationIntegration:
 # API Endpoint Tests
 # ============================================================================
 
+
 class TestQuestionValidationAPI:
     """Tests for question validation API endpoints."""
-    
+
     @pytest.mark.asyncio
-    async def test_validate_endpoint_returns_validation_result(self, valid_question_data):
+    async def test_validate_endpoint_returns_validation_result(
+        self, valid_question_data
+    ):
         """Test that the validate endpoint returns proper validation result."""
         from fastapi.testclient import TestClient
         from app.main import app
-        
+
         client = TestClient(app)
-        
+
         response = client.post(
-            "/api/question-validation/validate",
-            json=valid_question_data
+            "/api/question-validation/validate", json=valid_question_data
         )
-        
+
         assert response.status_code == 200
         data = response.json()
         assert "valid" in data
         assert "results" in data
-    
+
     @pytest.mark.asyncio
-    async def test_batch_validate_endpoint_validates_multiple_questions(self, valid_question_data):
+    async def test_batch_validate_endpoint_validates_multiple_questions(
+        self, valid_question_data
+    ):
         """Test that batch validate endpoint can validate multiple questions."""
         from fastapi.testclient import TestClient
         from app.main import app
-        
+
         client = TestClient(app)
-        
+
         questions = [valid_question_data, valid_question_data.copy()]
         questions[1]["id"] = "test-question-2"
-        
+
         response = client.post(
-            "/api/question-validation/batch-validate",
-            json=questions
+            "/api/question-validation/batch-validate", json=questions
         )
-        
+
         assert response.status_code == 200
         data = response.json()
         assert "results" in data
